// open_api_3.d.ts
export interface IPathItemObject {

  /**
   * Allows extensions to the OpenAPI Schema. The field name MUST begin with `x-`,
   * for example, `x-internal-id`. The value can be `null`, a primitive, an array or
   * an object. Can have any valid JSON format value.
   */
  [extension: string]: any;

  /**
   * Allows for an external definition of this path item. The referenced structure
   * MUST be in the format of a [Path Item Object](#pathItemObject). If there are
   * conflicts between the referenced definition and this Path Item's definition, the
   * behavior is undefined.
   */
  $ref: string;

  /**
   * An optional, string summary, intended to apply to all operations in this path.
   */
  summary: string;

  /**
   * An optional, string description, intended to apply to all operations in this
   * path. [CommonMark syntax](http://spec.commonmark.org/) MAY be used for rich text
   * representation.
   */
  description: string;

  /**
   * A definition of a GET operation on this path.
   */
  get: IOperationObject;

  /**
   * A definition of a PUT operation on this path.
   */
  put: IOperationObject;

  /**
   * A definition of a POST operation on this path.
   */
  post: IOperationObject;

  /**
   * A definition of a DELETE operation on this path.
   */
  delete: IOperationObject;

  /**
   * A definition of a OPTIONS operation on this path.
   */
  options: IOperationObject;

  /**
   * A definition of a HEAD operation on this path.
   */
  head: IOperationObject;

  /**
   * A definition of a PATCH operation on this path.
   */
  patch: IOperationObject;

  /**
   * A definition of a TRACE operation on this path.
   */
  trace: IOperationObject;

  /**
   * An alternative `server` array to service all operations in this path.
   */
  servers: Array<IServerObject>;

  /**
   * A list of parameters that are applicable for all the operations described under
   * this path. These parameters can be overridden at the operation level, but cannot
   * be removed there. The list MUST NOT include duplicated parameters. A unique
   * parameter is defined by a combination of a [name](#parameterName) and
   * [location](#parameterIn). The list can use the [Reference
   * Object](#referenceObject) to link to parameters that are defined at the [OpenAPI
   * Object's components/parameters](#componentsParameters).
   */
  parameters: Array<IParameterObject>;
}
// ...
export interface IOperationObject {

  /**
   * Allows extensions to the OpenAPI Schema. The field name MUST begin with `x-`,
   * for example, `x-internal-id`. The value can be `null`, a primitive, an array or
   * an object. Can have any valid JSON format value.
   */
  [extension: string]: any;

  /**
   * A list of tags for API documentation control. Tags can be used for logical
   * grouping of operations by resources or any other qualifier.
   */
  tags: Array<string>;

  /**
   * A short summary of what the operation does.
   */
  summary: string;

  /**
   * A verbose explanation of the operation behavior. [CommonMark
   * syntax](http://spec.commonmark.org/) MAY be used for rich text representation.
   */
  description: string;

  /**
   * Additional external documentation for this operation.
   */
  externalDocs: IExternalDocumentationObject;

  /**
   * Unique string used to identify the operation. The id MUST be unique among all
   * operations described in the API. Tools and libraries MAY use the operationId to
   * uniquely identify an operation, therefore, it is RECOMMENDED to follow common
   * programming naming conventions.
   */
  operationId: string;

  /**
   * A list of parameters that are applicable for this operation. If a parameter is
   * already defined at the [Path Item](#pathItemParameters), the new definition will
   * override it but can never remove it. The list MUST NOT include duplicated
   * parameters. A unique parameter is defined by a combination of a
   * [name](#parameterName) and [location](#parameterIn). The list can use the
   * [Reference Object](#referenceObject) to link to parameters that are defined at
   * the [OpenAPI Object's components/parameters](#componentsParameters).
   */
  parameters: Array<IParameterObject>;

  /**
   * The request body applicable for this operation.  The `requestBody` is only
   * supported in HTTP methods where the HTTP 1.1 specification
   * [RFC7231](https://tools.ietf.org/html/rfc7231#section-4.3.1) has explicitly
   * defined semantics for request bodies.  In other cases where the HTTP spec is
   * vague, `requestBody` SHALL be ignored by consumers.
   */
  requestBody: IRequestBodyObject;

  /**
   * The list of possible responses as they are returned from executing this
   * operation.
   */
  responses: IResponsesObject;

  /**
   * A map of possible out-of band callbacks related to the parent operation. The key
   * is a unique identifier for the Callback Object. Each value in the map is a
   * [Callback Object](#callbackObject) that describes a request that may be
   * initiated by the API provider and the expected responses. The key value used to
   * identify the callback object is an expression, evaluated at runtime, that
   * identifies a URL to use for the callback operation.
   */
  callbacks: { [name: string]: ICallbackObject };

  /**
   * Declares this operation to be deprecated. Consumers SHOULD refrain from usage of
   * the declared operation. Default value is `false`.
   */
  deprecated: boolean;

  /**
   * A declaration of which security mechanisms can be used for this operation. The
   * list of values includes alternative security requirement objects that can be
   * used. Only one of the security requirement objects need to be satisfied to
   * authorize a request. This definition overrides any declared top-level
   * [`security`](#oasSecurity). To remove a top-level security declaration, an empty
   * array can be used.
   */
  security: Array<ISecurityRequirementObject>;

  /**
   * An alternative `server` array to service this operation. If an alternative
   * `server` object is specified at the Path Item Object or Root level, it will be
   * overridden by this value.
   */
  servers: Array<IServerObject>;
}

/**
 * Allows referencing an external resource for extended documentation.
 * @example:
```json
  {
    "description": "Find more info here",
    "url": "https://example.com"
  }
```

```yaml
  description: Find more info here
  url: https://example.com
```
 */
export interface IExternalDocumentationObject {

  /**
   * Allows extensions to the OpenAPI Schema. The field name MUST begin with `x-`,
   * for example, `x-internal-id`. The value can be `null`, a primitive, an array or
   * an object. Can have any valid JSON format value.
   */
  [extension: string]: any;

  /**
   * A short description of the target documentation. [CommonMark
   * syntax](http://spec.commonmark.org/) MAY be used for rich text representation.
   */
  description: string;

  /**
   * The URL for the target documentation. Value MUST be in the format of a URL.
   */
  url: string;
}
// ...
export interface IParameterObject {

  /**
   * Allows extensions to the OpenAPI Schema. The field name MUST begin with `x-`,
   * for example, `x-internal-id`. The value can be `null`, a primitive, an array or
   * an object. Can have any valid JSON format value.
   */
  [extension: string]: any;

  /**
   * The name of the parameter. Parameter names are case sensitive. If
   * [`in`](#parameterIn) is `"path"`, the `name` field MUST correspond to the
   * associated path segment from the [path](#pathsPath) field in the [Paths
   * Object](#pathsObject). See [Path Templating](#pathTemplating) for further
   * information.If [`in`](#parameterIn) is `"header"` and the `name` field is
   * `"Accept"`, `"Content-Type"` or `"Authorization"`, the parameter definition
   * SHALL be ignored.For all other cases, the `name` corresponds to the parameter
   * name used by the [`in`](#parameterIn) property.
   */
  name: string;

  /**
   * The location of the parameter. Possible values are "query", "header", "path" or
   * "cookie".
   */
  in: string;

  /**
   * A brief description of the parameter. This could contain examples of use.
   * [CommonMark syntax](http://spec.commonmark.org/) MAY be used for rich text
   * representation.
   */
  description: string;

  /**
   * Determines whether this parameter is mandatory. If the [parameter
   * location](#parameterIn) is "path", this property is **REQUIRED** and its value
   * MUST be `true`. Otherwise, the property MAY be included and its default value is
   * `false`.
   */
  required: boolean;

  /**
   * Specifies that a parameter is deprecated and SHOULD be transitioned out of
   * usage.
   */
   deprecated: boolean;

  /**
   * Sets the ability to pass empty-valued parameters. This is valid only for `query`
   * parameters and allows sending a parameter with an empty value. Default value is
   * `false`. If [`style`](#parameterStyle) is used, and if behavior is `n/a` (cannot
   * be serialized), the value of `allowEmptyValue` SHALL be ignored.
   */
   allowEmptyValue: boolean;
}
// ...
export interface IRequestBodyObject {

  /**
   * Allows extensions to the OpenAPI Schema. The field name MUST begin with `x-`,
   * for example, `x-internal-id`. The value can be `null`, a primitive, an array or
   * an object. Can have any valid JSON format value.
   */
  [extension: string]: any;

  /**
   * A brief description of the request body. This could contain examples of use.
   * [CommonMark syntax](http://spec.commonmark.org/) MAY be used for rich text
   * representation.
   */
  description: string;

  /**
   * The content of the request body. The key is a media type or [media type
   * range](https://tools.ietf.org/html/rfc7231#appendix-D) and the value describes
   * it.  For requests that match multiple keys, only the most specific key is
   * applicable. e.g. text/plain overrides text/*
   */
  content: { [name: string]: IMediaTypeObject };

  /**
   * Determines if the request body is required in the request. Defaults to `false`.
   */
  required: boolean;
}
// ...
export interface IResponsesObject {

  /**
   * Any [HTTP status code](#httpCodes) can be used as the property name, but only
   * one property per code, to describe the expected response for that HTTP status
   * code.  A [Reference Object](#referenceObject) can link to a response that is
   * defined in the [OpenAPI Object's components/responses](#componentsResponses)
   * section. This field MUST be enclosed in quotation marks (for example, "200") for
   * compatibility between JSON and YAML. To define a range of response codes, this
   * field MAY contain the uppercase wildcard character `X`. For example, `2XX`
   * represents all response codes between `[200-299]`. The following range
   * definitions are allowed: `1XX`, `2XX`, `3XX`, `4XX`, and `5XX`. If a response
   * range is defined using an explicit code, the explicit code definition takes
   * precedence over the range definition for that code.
   * Allows extensions to the OpenAPI Schema. The field name MUST begin with `x-`,
   * for example, `x-internal-id`. The value can be `null`, a primitive, an array or
   * an object. Can have any valid JSON format value.
   */
  [httpStatusCodeHttpCodesAndExtension: string]: IResponseObject | any;

  /**
   * The documentation of responses other than the ones declared for specific HTTP
   * response codes. Use this field to cover undeclared responses. A [Reference
   * Object](#referenceObject) can link to a response that the [OpenAPI Object's
   * components/responses](#componentsResponses) section defines.
   */
  default: IResponseObject;
}
// ...
export interface ICallbackObject {

  /**
   * A Path Item Object used to define a callback request and expected responses.  A
   * [complete example](../examples/v3.0/callback-example.yaml) is available.
   * Allows extensions to the OpenAPI Schema. The field name MUST begin with `x-`,
   * for example, `x-internal-id`. The value can be `null`, a primitive, an array or
   * an object. Can have any valid JSON format value.
   */
  [expressionAndExtension: string]: IPathItemObject | any;
}